requirements.txt
Flask==2.0.3
gunicorn==20.1.0
Flask-Principal==0.4.0
Flask-WTF==1.0.0
Jinja2==3.0.3
MarkupSafe==2.0.1
WTForms==3.0.1
Werkzeug==2.0.3
blinker==1.4
click==8.0.3
flask-kafka==0.0.3
itsdangerous==2.0.1
kafka==1.3.5
kafka-python==1.3.5
kazoo==2.5.0
mysql-connector==2.2.9
pip==21.1.2
pykafka==2.8.0
setuptools==57.0.0
six==1.16.0
tabulate==0.8.9
wheel==0.36.2